fileservice: move some files to new memorycache package
